Added row and column containers
Added row and column containers.
Each container holds a list of all coords of all nodes on that x or y axis
Idea is to use these to reduce the net we cast looking for neighbors. We will only look at the nodes in the in the row and col containers of the node and its adjacent neighbors.

TODO:
Rewrite node creation function to support this new container data structure